requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
Flask-CORS==4.0.0

# Environment variables
python-dotenv==1.0.0

# Machine Learning
xgboost==1.7.6
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# AI and Vector Database
qdrant-client==1.14.2

# HTTP requests
requests==2.31.0

# Security and encryption
cryptography>=3.4.8
